refactor(registry): decompose ResumePane + my-jobs page under 200-line policy
ResumePane.jsx (295 -> 66): resume sections extracted into ResumeSections.jsx
(identity/skills/work) + ResumeSectionsExtra.jsx (education/languages/certs/
links); date + experience helpers into resumeUtils.js (+ test).

my-jobs/page.js (280 -> 138): JobCard.jsx + Filters.jsx pulled out; constants,
formatters, and filterJobs into jobFormatters.js (+ test). Both were honest
component boundaries, not layout churn. Behavior identical.
